+++ starsky/starskytest/starsky.foundation.accountmangement/Services/UserManagerTest.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
@@ -292,7 +293,27 @@ namespace starskytest.starsky.foundation.accountmangement.Services
 
 			Assert.AreEqual("GetUser",user.Name);
 		}
+		
+		[TestMethod]
+		public void GetUser_credentialTypeNull_IdDoesNotExist()
+		{
+			var userManager = new UserManager(_dbContext, new AppSettings(), _memoryCache);
+			var user = userManager.GetUser("email", "sfkknfdlknsdfl@mail.us");
+			Assert.IsNull(user);
+		}
 
+				
+		[TestMethod]
+		public void GetUser_IdDoesNotExist()
+		{
+			var userManager = new UserManager(_dbContext, new AppSettings(), _memoryCache);
+			userManager.AddDefaultCredentialType("email");
+			
+			var user = userManager.GetUser("email", "sfkknfdlknsdfl@mail.us");
+			Assert.IsNull(user);
+		}
+		
+		
 		[TestMethod]
 		public void PreflightValidate_Fail_stringEmpty()
 		{
@@ -371,6 +392,52 @@ namespace starskytest.starsky.foundation.accountmangement.Services
 
 			Assert.AreEqual(1, result.Count);
 			Assert.AreEqual("test", result[0].Value);
+		}
+		
+				
+		[TestMethod]
+		public async Task Cache_ExistsByUserTableId_HitResult()
+		{
+			var userManager = new UserManager(_dbContext,new AppSettings(), _memoryCache);
+			await userManager.AddUserToCache(new User{Name = "cachedUser", Id = 1});
+
+			var result = await userManager.Exist(1);
+			
+			Assert.IsNotNull(result);
+		}
+		
+		[TestMethod]
+		public async Task Cache_ExistsByUserTableId_NotFound()
+		{
+			var userManager = new UserManager(_dbContext,new AppSettings(), _memoryCache);
+			await userManager.AddUserToCache(new User{Name = "cachedUser", Id = 1});
+
+			var result = await userManager.Exist(9822);
+			
+			Assert.IsNull(result);
+		}
+		
+						
+		[TestMethod]
+		public async Task Db_ExistsByUserTableId_HitResult()
+		{
+			var userManager = new UserManager(_dbContext,new AppSettings{AddMemoryCache = false}, _memoryCache);
+			var id = await userManager.SignUpAsync(string.Empty, "email", "t", "t");
+			Assert.IsNotNull(id);
+
+			var result = await userManager.Exist(id.User.Id);
+			
+			Assert.IsNotNull(result);
+		}
+		
+		[TestMethod]
+		public async Task Db_ExistsByUserTableId_NotFound()
+		{
+			var userManager = new UserManager(_dbContext,new AppSettings{AddMemoryCache = false}, _memoryCache);
+
+			var result = await userManager.Exist(852);
+			
+			Assert.IsNull(result);
 		}
 	}
 }